[feat #109] 관리자 권한 여부 확인
