+++ src/main/java/com/umc/ttt/domain/bookLetter/converter/BookLetterConverter.java
@@ -8,6 +8,7 @@ import com.umc.ttt.domain.bookLetter.dto.BookLetterResponseDTO;
 import com.umc.ttt.domain.bookLetter.entity.BookLetter;
 import com.umc.ttt.domain.bookLetter.entity.BookLetterBook;
 import com.umc.ttt.domain.bookLetter.entity.BookLetterCategory;
+import com.umc.ttt.domain.member.entity.Member;
 import org.springframework.data.domain.Page;
 
 import java.util.List;
@@ -20,13 +21,14 @@ public class BookLetterConverter {
                 .build();
     }
 
-    public static BookLetter toBookLetter(BookLetterRequestDTO.CUDto request, String coverImg) {
+    public static BookLetter toBookLetter(BookLetterRequestDTO.CUDto request, String coverImg, Member member) {
         return BookLetter.builder()
                 .title(request.getTitle())
                 .subtitle(request.getSubtitle())
                 .editor(request.getEditor())
                 .content(request.getContent())
                 .coverImg(coverImg)
+                .editorMember(member)
                 .build();
     }
 
@@ -98,15 +100,17 @@ public class BookLetterConverter {
                 .build();
     }
 
-    public static BookLetterResponseDTO.BookLetterDTO bookLetterDTO(BookLetter bookLetter) {
+    public static BookLetterResponseDTO.BookLetterDTO bookLetterDTO(BookLetter bookLetter, Member member) {
         List<BookLetterResponseDTO.BookDTO> bookDTOList = bookLetter.getBooks().stream()
                 .map(BookLetterConverter::bookDTO).collect(Collectors.toList());
+
+        boolean isWriter = bookLetter.getEditorMember().equals(member);
 
         return BookLetterResponseDTO.BookLetterDTO.builder()
                 .title(bookLetter.getTitle())
                 .subtitle(bookLetter.getSubtitle())
                 .editor(bookLetter.getEditor())
-                .isWriter(true) // 추후 관리자 여부 확인 로직되면 수정
+                .isWriter(isWriter)
                 .content(bookLetter.getContent())
                 .coverImg(bookLetter.getCoverImg())
                 .bookList(bookDTOList)
